Add tile module
